Squashed commit of the following:
commit af4a6d23b94bf71320e7300252aeadb8c03cd299
Author: LeThienTrungNguyen <khamnguyenakaams2001@gmail.com>
Date:   Fri Nov 17 21:42:23 2023 +0700

    Update Sword.prefab

    1

commit 0d6875180e0cc59c5ff8fa3a2e7fbf28c8a356a8
Author: LeThienTrungNguyen <khamnguyenakaams2001@gmail.com>
Date:   Fri Nov 17 21:36:06 2023 +0700

    1

    1

commit f918a877f9d6d5d166299917d5b22f71437d489b
Author: LeThienTrungNguyen <khamnguyenakaams2001@gmail.com>
Date:   Thu Nov 16 21:25:20 2023 +0700

    1

    1

commit 79943c7325f1011c344263ef745315ecb5370aaa
Merge: a90e301 b1fe642
Author: LeThienTrungNguyen <khamnguyenakaams2001@gmail.com>
Date:   Wed Nov 15 23:51:38 2023 +0700

    Merge branch 'main' into nguyen

commit a90e301d18e3c9ea2918ad055c44f7b6c204d9a9
Author: LeThienTrungNguyen <khamnguyenakaams2001@gmail.com>
Date:   Tue Nov 14 22:32:26 2023 +0700

    1

    1

commit 8ef166b280c6ea9a0a02bb40ac72fb526e67b61d
Author: LeThienTrungNguyen <khamnguyenakaams2001@gmail.com>
Date:   Tue Nov 14 02:16:58 2023 +0700

    1

    1

commit a6d4fb915f74028ecf62920d86406fc9ae0db8c1
Author: LeThienTrungNguyen <khamnguyenakaams2001@gmail.com>
Date:   Tue Nov 14 00:53:58 2023 +0700

    update indicate for boss final

commit 4701e04671e152a57ce42e49af78676ef05a9e1c
Author: LeThienTrungNguyen <khamnguyenakaams2001@gmail.com>
Date:   Mon Nov 13 18:10:25 2023 +0700

    update collider weapon

commit 0ee1a06159d3a8046208795992af1d645b491c97
Author: LeThienTrungNguyen <khamnguyenakaams2001@gmail.com>
Date:   Mon Nov 13 18:06:58 2023 +0700

    add final skill for boss final

commit 956ff5e5281bab8e22ad0904c3879ba5361dd73a
Author: LeThienTrungNguyen <khamnguyenakaams2001@gmail.com>
Date:   Sat Nov 11 23:50:33 2023 +0700

    update and add boss skill

commit ebcc3d4c430028b99fa5f61c5c942db13b8ed905
Author: LeThienTrungNguyen <khamnguyenakaams2001@gmail.com>
Date:   Fri Nov 10 02:07:23 2023 +0700

    add boss final phase 3

commit 1a99fd3d93a56a1c42d8afa8038bf77413e66e21
Author: LeThienTrungNguyen <khamnguyenakaams2001@gmail.com>
Date:   Wed Nov 8 22:28:42 2023 +0700

    boss 1 particle update

    1

commit 2d80856c1de4eeda6be8a1e5a44d10437e994f21
Author: LeThienTrungNguyen <khamnguyenakaams2001@gmail.com>
Date:   Mon Nov 6 22:41:14 2023 +0700

    1

    1

commit f05e76fd8557d51f766f1772ac72ecd75d81dd90
Author: LeThienTrungNguyen <khamnguyenakaams2001@gmail.com>
Date:   Mon Nov 6 22:29:59 2023 +0700

    complete boss lv1

commit 8bb1ab3d58735b64d9b54cd9dc5a540f4cb566b8
Author: LeThienTrungNguyen <khamnguyenakaams2001@gmail.com>
Date:   Mon Nov 6 00:14:02 2023 +0700

    test commit by nguyen

commit c7d5f909afd00bf99cef1b5fdabef4130a34abb9
Author: LeThienTrungNguyen <112458785+LeThienTrungNguyen@users.noreply.github.com>
Date:   Sun Nov 5 23:15:56 2023 +0700

    1

    1

commit a3da2419f9daf2bf60071b189b41c7be906711e5
Author: LeThienTrungNguyen <112458785+LeThienTrungNguyen@users.noreply.github.com>
Date:   Sun Nov 5 15:48:05 2023 +0700

    1

    1

commit 920eacf817b347ffa811d22f06b01304566f1698
Author: LeThienTrungNguyen <112458785+LeThienTrungNguyen@users.noreply.github.com>
Date:   Sun Nov 5 15:47:32 2023 +0700

    1

    1

commit f6518dd5158db428dbbe809a1d730a5903ee4606
Author: LeThienTrungNguyen <112458785+LeThienTrungNguyen@users.noreply.github.com>
Date:   Sun Nov 5 15:27:52 2023 +0700

    add boss lv1 skill

    phase 1:
    + rush hit wall
    + melee attack
    + jump to player

    Phase 2 :
    + rush hit wall
    + ground shake

commit 6adb034567becdf51071ff6e79456b0263e1b3a4
Author: LeThienTrungNguyen <112458785+LeThienTrungNguyen@users.noreply.github.com>
Date:   Sat Nov 4 23:54:09 2023 +0700

    add boss lv1 skill

    rush hit wall
    move
    evade

+++ Assets/NguyenFolder/Script/Boss_Lv1_Fake_Jump.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 
 public class Boss_Lv1_Fake_Jump : StateMachineBehaviour
 {
-    public float numerator;
-    public float denominator;
+    public int numerator;
+    public int denominator;
 
     public float heightOffset;
     float fraction;
